Bump iOS CQ bots to macOS 10.11.
ios64_compile does not seem to work on 10.9, probably because the clang
integrated assembler was still missing some ARM features at the time.
https://luci-logdog.appspot.com/v/?s=boringssl%2Fbuildbucket%2Fcr-buildbucket.appspot.com%2F8979949146687558800%2F%2B%2Fsteps%2Fninja%2F0%2Fstdout

I'm told Chromium for iOS uses something approximating the 10.11 SDK, so
let's use that for the CQ. Chromium still supports 10.9 (for running,
though the toolchain is newer) for macOS, so I've left those alone.

Change-Id: Ibe8dccab83617c04a5320b5f4d1ec79e1190436f
Reviewed-on: https://boringssl-review.googlesource.com/16150
Reviewed-by: Andrii Shyshkalov <tandrii@google.com>